requirements.txt
flask
flask_cors
requests
tweepy
Pillow
keyring